requirements.txt
asgiref==3.5.2
astroid==2.12.13
autopep8==2.0.0
colorama==0.4.6
coverage==6.5.0
dill==0.3.6
dj-database-url==1.0.0
Django==4.1.3
django-cors-headers==3.10.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
drf-flex-fields==1.0.0
greenlet==2.0.1
gunicorn==20.1.0
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
pep8==1.7.1
platformdirs==2.5.4
psycopg2-binary==2.9.5
psycopg2==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
pylint==2.15.6
pylint-django==2.5.3
pylint-plugin-utils==0.7
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
tzdata==2022.6
whitenoise==6.2.0
wrapt==1.14.1
xlsxwriter==3.0.3
dj_database_url==1.0.0
psycopg2==2.9.5
whitenoise==6.2.0
requests==2.28.1